+++ MediaContentSynchronizationApi/etc/di.xml
@@ -6,5 +6,5 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\MediaContentSynchronizationApi\Model\GetEntitiesInterface" type="Magento\MediaContentSynchronizationApi\Model\Entities"/>
+    <preference for="Magento\MediaContentSynchronizationApi\Model\GetEntitiesInterface" type="Magento\MediaContentSynchronizationApi\Model\GetEntities"/>
 </config>
